HUGE REFACTOR (see description)
- BREAKING CHANGE : TIME IS NOW EXPRESSED IN SECONDS
- Moved ModelTimer and IUpdateable
- Removed IMovementHandler and MovementHandler isn't generic anymore
- A lot of refactor and renaming everywhere

+++ src/main/java/model/ingame/weapon/AbstractTrapPlacer.java
@@ -7,7 +7,7 @@ import model.ingame.entity.ICombatEntity;
 public abstract class AbstractTrapPlacer extends WeaponModel {
     protected int trapCount;
 
-    public AbstractTrapPlacer(String name, String identifier, GameModel gameModel, ICombatEntity owner, int coolDown, int trapCount) {
+    public AbstractTrapPlacer(String name, String identifier, GameModel gameModel, ICombatEntity owner, double coolDown, int trapCount) {
         super(name, identifier, gameModel, owner, coolDown);
         this.trapCount = trapCount;
     }
